requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
passlib[bcrypt]
python-multipart
email-validator
openai
gunicorn
python-dotenv
psycopg2-binary
